// backend/src/activity-tracker/enums/activity-action.enum.ts
export enum ActivityAction {
  LOGIN = "login",
  LOGOUT = "logout",
  DOCUMENT_UPLOAD = "document_upload",
  DOCUMENT_VIEW = "document_view",
  DOCUMENT_DOWNLOAD = "document_download",
  DOCUMENT_UPDATE = "document_update",
  DOCUMENT_DELETE = "document_delete",
  DOCUMENT_VERSION_VIEW = "document_version_view",
  FEEDBACK_SUBMIT = "feedback_submit",
  FEEDBACK_UPDATE = "feedback_update",
  FEEDBACK_COMMENT = "feedback_comment",
  TAG_CREATE = "tag_create",
  TAG_UPDATE = "tag_update",
  TAG_DELETE = "tag_delete",
  OWNER_CREATE = "owner_create",
  OWNER_UPDATE = "owner_update",
  OWNER_DELETE = "owner_delete",
  USER_PROFILE_UPDATE = "user_profile_update",
  PASSWORD_CHANGE = "password_change",
  SYSTEM_ADMIN_ACTION = "system_admin_action",
  OTHER = "other",
}